Add LICENSE and README.md files
This commit adds a LICENSE file to the repository, which specifies the terms under which the code can be used, modified, and distributed. It also adds a README.md file, which provides an overview of the `sudoku` package, instructions for installation and usage, and information about licensing. These changes improve the usability and accessibility of the package.