Add support for reading puzzles from text file
This commit adds a new `parse_puzzle` function to the `sudoku/parse_sudoku.py` module, which can be used to read a Sudoku puzzle from standard input. The `cli.py` module has been updated to use this function if no puzzle file is specified on the command line. This allows users to enter a puzzle directly into the terminal, rather than having to create a separate file. The `__init__.py` file has been updated to export the `parse_puzzle` function. These changes improve the usability and flexibility of the `sudoku` package.